Make DateRecurringIndexQueryParser optional in order to
maintain compatibility with vanilla Plone 4.3.
